chore : Wizard 클래스 새로 생성
